Refactored “drive with joystick” into parts
I changed `driveWithJoystick` so that it instead of being a single function that reads the joystick, scales the inputs, and activates the swerve, it now does that in 3 steps: `getJoystickDriveValues`, `speeds_for_joystick_values`, and then the call to `swerve.drive`. By separting input, calculation, and output, we can test and debug better.